Do not require "sidekiq/cron" to run "sidekiq/cron/web"
- Extract config into Sidekiq::Cron::Config
- Run isolated test for web_extension_test.rb
