+++ .github/workflows/ci.yml
@@ -31,5 +31,9 @@ jobs:
         ruby-version: ${{ matrix.ruby }}
     - name: Run tests
       run: bundle exec rake test
+    - name: Run web_extension_test.rb without sidekiq/cron require
+      env:
+        TEST: test/unit/web_extension_test.rb
+      run:  bundle exec rake test
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
